sql/catalog/descs: move some code into new files
Preliminary before more refactoring.

Release note: None
