+++ pkg/sql/catalog/descs/leased_descriptors.go
@@ -1,0 +1,90 @@
+// Copyright 2021 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package descs
+
+import (
+	"sort"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/lease"
+)
+
+// leasedDescriptors holds references to all the descriptors leased in the
+// transaction, and supports access by name and by ID.
+type leasedDescriptors struct {
+	descs []lease.LeasedDescriptor
+}
+
+func (ld *leasedDescriptors) add(desc lease.LeasedDescriptor) {
+	ld.descs = append(ld.descs, desc)
+}
+
+func (ld *leasedDescriptors) releaseAll() (toRelease []lease.LeasedDescriptor) {
+	toRelease = append(toRelease, ld.descs...)
+	ld.descs = ld.descs[:0]
+	return toRelease
+}
+
+func (ld *leasedDescriptors) release(ids []descpb.ID) (toRelease []lease.LeasedDescriptor) {
+	// Sort the descriptors and leases to make it easy to find the leases to release.
+	leasedDescs := ld.descs
+	sort.Slice(ids, func(i, j int) bool {
+		return ids[i] < ids[j]
+	})
+	sort.Slice(leasedDescs, func(i, j int) bool {
+		return leasedDescs[i].Desc().GetID() < leasedDescs[j].Desc().GetID()
+	})
+
+	filteredLeases := leasedDescs[:0] // will store the remaining leases
+	idsToConsider := ids
+	shouldRelease := func(id descpb.ID) (found bool) {
+		for len(idsToConsider) > 0 && idsToConsider[0] < id {
+			idsToConsider = idsToConsider[1:]
+		}
+		return len(idsToConsider) > 0 && idsToConsider[0] == id
+	}
+	for _, l := range leasedDescs {
+		if !shouldRelease(l.Desc().GetID()) {
+			filteredLeases = append(filteredLeases, l)
+		} else {
+			toRelease = append(toRelease, l)
+		}
+	}
+	ld.descs = filteredLeases
+	return toRelease
+}
+
+func (ld *leasedDescriptors) getByID(id descpb.ID) catalog.Descriptor {
+	for i := range ld.descs {
+		desc := ld.descs[i]
+		if desc.Desc().GetID() == id {
+			return desc.Desc()
+		}
+	}
+	return nil
+}
+
+func (ld *leasedDescriptors) getByName(
+	dbID descpb.ID, schemaID descpb.ID, name string,
+) catalog.Descriptor {
+	for i := range ld.descs {
+		desc := ld.descs[i]
+		if lease.NameMatchesDescriptor(desc.Desc(), dbID, schemaID, name) {
+			return desc.Desc()
+		}
+	}
+	return nil
+}
+
+func (ld *leasedDescriptors) numDescriptors() int {
+	return len(ld.descs)
+}
